[backend] Playbook : added filter to access restriction component (#14285)
