+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/playbookComponents/access-restrictions-component-test.ts
@@ -8,10 +8,12 @@ import { testBundleObject, testExecutor } from './playbook-components-test-utils
 import { ENTITY_TYPE_CONTAINER_GROUPING } from '../../../../src/modules/grouping/grouping-types';
 import { USER_EDITOR } from '../../../utils/testQuery';
 import { playbookBundleElementsToApply } from '../../../../src/modules/playbook/playbook-types';
+import { ENTITY_TYPE_CONTAINER_CASE_INCIDENT } from '../../../../src/modules/case/case-incident/case-incident-types';
 
 describe('PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT', () => {
   const REPORT_ID = 'report--5f78a68b-2c4d-5e6f-beaa-7b987b0e7165';
   const GROUPING_ID = 'grouping--5f78a68b-2c4d-5e6f-beaa-7b987b0e7133';
+  const CASE_INCIDENT_ID = 'case-incident--5f78a68b-2c4d-5e6f-beaa-7b987b0e7145';
   const USER_ID = USER_EDITOR.id;
   const ORGA_ID = 'org-uuid-1';
   const GROUP_ID = 'group-uuid-1';
@@ -404,7 +406,7 @@ describe('PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT', () => {
       }),
     ];
 
-    it('should add authorized_members to all objects in the bundle', async () => {
+    it('should add authorized_members to all objects in the bundle when applyToElements = "all-element"', async () => {
       const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
         mainId: REPORT_ID,
         bundleObjects: BUNDLE_OBJECTS(),
@@ -423,7 +425,7 @@ describe('PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT', () => {
       expect(secondReportExt.authorized_members).toEqual(expectedAccessRestriction);
     });
 
-    it('should only add authorized_members to only main element of the bundle', async () => {
+    it('should only add authorized_members to only main element of the bundle when applyToElements = "only-main"', async () => {
       const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
         mainId: REPORT_ID,
         bundleObjects: BUNDLE_OBJECTS(),
@@ -442,7 +444,7 @@ describe('PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT', () => {
       expect(secondReportExt.authorized_members).toBeUndefined();
     });
 
-    it('should add authorized_members to all except main element of the bundle', async () => {
+    it('should add authorized_members to all except main element of the bundle when applyToElements = "all-except-main"', async () => {
       const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
         mainId: REPORT_ID,
         bundleObjects: BUNDLE_OBJECTS(),
@@ -455,11 +457,223 @@ describe('PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT', () => {
       expect(result.output_port).toBe('out');
       // Check first object - should not have authorized_members
       const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
-      console.log();
       expect(reportExt.authorized_members).toBeUndefined();
       // Check second object - should have authorized_members
       const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
       expect(secondReportExt.authorized_members).toEqual(expectedAccessRestriction);
+    });
+  });
+
+  describe('when using filters on bundle containing multiple objects', () => {
+    const filterGrouping = '{"mode":"and","filters":[{"key":["entity_type"],"operator":"eq","values":["Grouping"],"mode":"or"}],"filterGroups":[]}';
+    const filterReport = '{"mode":"and","filters":[{"key":["entity_type"],"operator":"eq","values":["Report"],"mode":"or"}],"filterGroups":[]}';
+    const filterAll = '{"mode":"and","filters":[{"key":["entity_type"],"operator":"eq","values":["Report", "Grouping", "Case-Incident"],"mode":"or"}],"filterGroups":[]}';
+    const filterReportAndGrouping = '{"mode":"and","filters":[{"key":["entity_type"],"operator":"eq","values":["Report", "Grouping"],"mode":"or"}],"filterGroups":[]}';
+    const filterNotMatching = '{"mode":"and","filters":[{"key":["entity_type"],"operator":"eq","values":["Incident", "Indicator"],"mode":"or"}],"filterGroups":[]}';
+
+    const expectedAccessRestriction = [{
+      id: USER_ID,
+      access_right: 'admin',
+      groups_restriction_ids: [],
+    }];
+
+    const BUNDLE_OBJECTS = () => [
+      testBundleObject({
+        id: REPORT_ID,
+        type: ENTITY_TYPE_CONTAINER_REPORT,
+      }),
+      testBundleObject({
+        id: GROUPING_ID,
+        type: ENTITY_TYPE_CONTAINER_GROUPING,
+      }),
+      testBundleObject({
+        id: CASE_INCIDENT_ID,
+        type: ENTITY_TYPE_CONTAINER_CASE_INCIDENT,
+      }),
+    ];
+
+    it('should only add authorized_members to only main element of the bundle when applyToElements = "only-main" and filter on all elements', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.onlyMain.value,
+          applyWithFilters: filterAll,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toEqual(expectedAccessRestriction);
+      // Check second object - should not have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toBeUndefined();
+      // Check third object - should not have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toBeUndefined();
+    });
+
+    it('should not add authorized_members to any element of the bundle when applyToElements = "only-main" and filter do not match main', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.onlyMain.value,
+          applyWithFilters: filterGrouping,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should not have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toBeUndefined();
+      // Check second object - should not have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toBeUndefined();
+      // Check third object - should not have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toBeUndefined();
+    });
+
+    it('should add authorized_members to all elements when filtered match on all and applyToElements = "all-element"', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.allElements.value,
+          applyWithFilters: filterAll,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toEqual(expectedAccessRestriction);
+      // Check second object - should have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toEqual(expectedAccessRestriction);
+      // Check third object - should have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toEqual(expectedAccessRestriction);
+    });
+
+    it('should add authorized_members only to groupings when filtered on groupings and applyToElements = "all-element"', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.allElements.value,
+          applyWithFilters: filterGrouping,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should not have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toBeUndefined();
+      // Check second object - should have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toEqual(expectedAccessRestriction);
+      // Check third object - should not have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toBeUndefined();
+    });
+
+    it('should not add authorized_members to any element when filters do not match and applyToElements = "all-element"', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.allElements.value,
+          applyWithFilters: filterNotMatching,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should not have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toBeUndefined();
+      // Check second object - should not have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toBeUndefined();
+      // Check third object - should not have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toBeUndefined();
+    });
+
+    it('should not add authorized_members to any element if applyToElements = "all-except-main" and filter is on main', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.allExceptMain.value,
+          applyWithFilters: filterReport,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should not have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toBeUndefined();
+      // Check second object - should not have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toBeUndefined();
+      // Check third object - should not have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toBeUndefined();
+    });
+
+    it('should add authorized_members filtered element except main if applyToElements = "all-except-main" and filter matches main and other elements', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.allExceptMain.value,
+          applyWithFilters: filterReportAndGrouping,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should not have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toBeUndefined();
+      // Check second object - should have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toEqual(expectedAccessRestriction);
+      // Check third object - should not have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toBeUndefined();
+    });
+
+    it('should add authorized_members to all elements except main if applyToElements = "all-except-main" and filter match all elements', async () => {
+      const result = await PLAYBOOK_ACCESS_RESTRICTIONS_COMPONENT.executor(testExecutor({
+        mainId: REPORT_ID,
+        bundleObjects: BUNDLE_OBJECTS(),
+        configuration: {
+          access_restrictions: [createAccessRestriction(USER_ID, 'admin')],
+          applyToElements: playbookBundleElementsToApply.allExceptMain.value,
+          applyWithFilters: filterAll,
+        },
+      }));
+
+      expect(result.output_port).toBe('out');
+      // Check first object - should not have authorized_members
+      const reportExt = result.bundle.objects[0].extensions[STIX_EXT_OCTI];
+      expect(reportExt.authorized_members).toBeUndefined();
+      // Check second object - should have authorized_members
+      const secondReportExt = result.bundle.objects[1].extensions[STIX_EXT_OCTI];
+      expect(secondReportExt.authorized_members).toEqual(expectedAccessRestriction);
+      // Check third object - should have authorized_members
+      const thirdReportExt = result.bundle.objects[2].extensions[STIX_EXT_OCTI];
+      expect(thirdReportExt.authorized_members).toEqual(expectedAccessRestriction);
     });
   });
